ot exe automake support! Windows needs a small amount of work.
