+++ include/otlib/OTScript.h
@@ -209,7 +209,7 @@ public:
     //
     void AddParty       (const std::string str_party_name,    OTParty & theParty);
     void AddAccount     (const std::string str_acct_name,     OTPartyAccount & theAcct);
-    void AddVariable    (const std::string str_var_name,      OTVariable & theVar);
+EXPORT    void AddVariable    (const std::string str_var_name,      OTVariable & theVar);
     
     // Note: any relevant assets or asset accounts are listed by their owner / contributor
     // parties. Therefore there's no need to separately input any accounts or assets to
@@ -223,7 +223,7 @@ typedef std::auto_ptr<OTScript>         OTScript_AutoPtr;
 
 // -----------------------------------
 
-OTScript_AutoPtr OTScriptFactory(const std::string & script_contents, 
+EXPORT OTScript_AutoPtr OTScriptFactory(const std::string & script_contents, 
                                  const std::string * p_script_type=NULL);
