+++ project/_3rdparty/zeromq2-x_msvc.7z
